+++ src/routes/artist.tsx
@@ -18,6 +18,7 @@ import {
 } from "../stores/video";
 import { searchYouTubeVideo } from "../api/youtube";
 import useSessionStore from "../stores/session";
+import EllipsisMenu from "../components/ellipsisMenu";
 
 export default function Artist() {
   const [isLoading, setIsLoading] = useState<boolean>(true);
@@ -105,7 +106,7 @@ export default function Artist() {
         return (
           <div
             key={item.id}
-            className="py-1 hover:bg-orange-200 hover:rounded-md grid grid-cols-[1fr_20fr_1fr] group mr-5"
+            className="py-1 hover:bg-orange-200 hover:rounded-md grid grid-cols-[1fr_20fr_1fr] group mr-5 relative"
           >
             <div className="group-hover:hidden flex justify-center items-center">
               {index + 1}
@@ -137,9 +138,12 @@ export default function Artist() {
                 })}
               </div>
             </div>
-            <div className="flex justify-center items-center">
-              {duration_min}:{duration_sec}
-            </div>
+            <EllipsisMenu
+              id={item.id}
+              name={item.name}
+              artists={artists}
+              imgUrl={item.album.images[0].url}
+            />
           </div>
         );
       })}
